Stopped caching fulfills calls.
I don't think it's legitimate to cache calls to fulfills, since it's likely to be called with lots of instances, potentially short lived, and caching those calls is likely to be a memory leak issue.  The best we can probably do is make a caching call to isA when the protocol is an interface.