Clear assignable cache whenever target changes
This fix invalidates the `__assignable_from_cache__` cache on target,
whenever
you call a function that modifies it (extend, inherit, ...).

The original bug was that if you called topiarist.inherit() it would
internally
call `isClassA()` which would then cache the result as "no this isn't
A", but
then the inherit would actually make the target an "A". And then if you
later
called `isClassA` it would come back as `false` even though it actually
is an
"A".