Demo config to put source code into package root
This commit shows the practical effects of placing source code into the
root of the package insead of a `src` directory. Specifically, the
`pretest:lint` and `test:browser` commands need to explicitly mention
each source file and directory.

It would be good to repeat this experiment once more commands have been
added, for example:

  1. Code-style checking.
  2. Static type checking.
  3. Code coverage generation.